fixup! fix(fetchMap): Support more clusterTile layer props
